[fix]refactor main and separate utils for nn_train and preprocess
